remove example.ts
